mcp_server/requirements.txt

fastmcp
supabase
python-dotenv
